Classifyed ReadPairer
