+++ milo/tests.py
@@ -1,9 +1,9 @@
 import unittest
 
 from genomicsUtils import *
-from pairedFastqProc import *
+from AmpliconMatcherProbabilistic import *
 from AmpliconMatcherHashSweep import *
-
+from ReadPairer import *
 
 # Here's our "unit tests".
 class GenomicsUtils(unittest.TestCase):
@@ -83,27 +83,30 @@ class AmpliconMatcherHashSweepTests(unittest.TestCase):
 
 
 
-class PairedFASTQAligher(unittest.TestCase):
+class PairedFASTQAligner(unittest.TestCase):
+
+    def setUp(self):
+        self.readPairer = ReadPairer()
 
     def test_scoring(self):
         left = "AAAAAAAAAAAAA"
         right = "AAAAAAAAAAAAA"
         overlapPairs = tuple(zip(left, right))
-        score = calcScore(overlapPairs, len(left))
+        score = self.readPairer.calcScore(overlapPairs, len(left))
         self.assertEqual(score, 13)
 
     def test_score_ignore(self):
         left = "AAAAAAAAAAAAA"
         right = "CCCCCCCCCCCCC"
         overlapPairs = tuple(zip(left, right))
-        score = calcScore(overlapPairs, len(left))
+        score = self.readPairer.calcScore(overlapPairs, len(left))
         self.assertEqual(score, -5)
 
     def test_score_mismatch(self):
         left = "AAAAAAAAAAAAA"
         right = "AAAAAAAAAAAAC"
         overlapPairs = tuple(zip(left, right))
-        score = calcScore(overlapPairs, len(left))
+        score = self.readPairer.calcScore(overlapPairs, len(left))
         self.assertEqual(score, 7)
 
     def test_no_overlap(self):
@@ -112,7 +115,7 @@ class PairedFASTQAligher(unittest.TestCase):
         leftQuality = "K" * len(left)
         rightQuality = "K" * len(right)
 
-        mergedSequence, qualityScores, noOfCol = mergeUnpaired(left, right, leftQuality, rightQuality)
+        mergedSequence, qualityScores, noOfCol = self.readPairer.mergeUnpaired(left, right, leftQuality, rightQuality)
         expected = left + " " + right
 
         # No overlap because too small
@@ -127,7 +130,7 @@ class PairedFASTQAligher(unittest.TestCase):
         leftQuality = "K" * len(left)
         rightQuality = "K" * len(right)
 
-        mergedSequence, qualityScores, noOfCol = mergeUnpaired(left, right, leftQuality, rightQuality)
+        mergedSequence, qualityScores, noOfCol = self.readPairer.mergeUnpaired(left, right, leftQuality, rightQuality)
 
         expected = left + " " + right
 
@@ -143,7 +146,7 @@ class PairedFASTQAligher(unittest.TestCase):
         leftQuality = "K" * len(left)
         rightQuality = "K" * len(right)
 
-        mergedSequence, qualityScores, noOfCol = mergeUnpaired(left, right, leftQuality, rightQuality)
+        mergedSequence, qualityScores, noOfCol = self.readPairer.mergeUnpaired(left, right, leftQuality, rightQuality)
 
         expected = "A" * sides + "C" * overlap + "G" * sides
 
@@ -157,7 +160,7 @@ class PairedFASTQAligher(unittest.TestCase):
         leftQuality = "K" * len(left)
         rightQuality = "K" * len(right)
 
-        mergedSequence, qualityScores, noOfCol = mergeUnpaired(left, right, leftQuality, rightQuality)
+        mergedSequence, qualityScores, noOfCol = self.readPairer.mergeUnpaired(left, right, leftQuality, rightQuality)
 
         expected = "A" * sides + "C" * overlap + "G" * sides
 
@@ -169,7 +172,7 @@ class PairedFASTQAligher(unittest.TestCase):
         right = reverseComplement("AGTCAGGATGTTAGCAGAGCCAGTCAAGACTTGCCGACAAAGGAAACTAGAAGCCAAGAAAGCTGCAGCTGAAAAGCTTTCCTCCCTGGAGAACAGCTCAAATAAAAATGAAAAGGAAAAGTCAGCCCCATCACGTACAAAACAAACTGAA")
         lQuality = "CCCCCGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGFGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGG"
         rQuality = "CCCCCGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGFGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGDGGGGGGGGGGGGGGGGGGGGFGGGGGGGGGGGDFG"[::-1]
-        mergedSequence, qualityScores, noOfCol = mergeUnpaired(left, right, lQuality, rQuality)
+        mergedSequence, qualityScores, noOfCol = self.readPairer.mergeUnpaired(left, right, lQuality, rQuality)
 
         expectedMerged = "CAAGTTGATGGGGGCAAAACCAAAATAATTTTCATTTTTAATATACCACACAACACATCTATCTACAAATGCTTTACATTAAATTTACCTGCCAACTGTTTAGCCTGGCTTGCGTTTTCAGTTTGTTTTGTACGTGATGGGGCTGACTTTTCCTTTTCATTTTTATTTGAGCTGTTCTCCAGGGAGGAAAGCTTTTCAGCTGCAGCTTTCTTGGCTTCTAGTTTCCTTTGTCGGCAAGTCTTGACTGGCTCTGCTAACATCCTGACT"
         expecedQuality = "." * len(expectedMerged)
@@ -184,7 +187,7 @@ class PairedFASTQAligher(unittest.TestCase):
         lQuality = "CCCCCGGGFCCEF;F9AGFGFDEGGGGGDFAAEFF<FG<CFFFFGFEG8F@C@C7<C:FC7FD8<FDCAC<FEGGC,CFGACFFGGG<F6C<C9<6C@@FF@CCFFGEGGGGGGGGFG8=?4EEFFFCCCFGGFFF<EGGGGCF8?==BC@"
         rQuality = "-8BC8B@FC<9E--6FFECEFGGG8CCFCGGG,CFFF,,,;EFFCCC,CFF6CFCA8,CC,C96C@EF9,C,,66,<,E<,@FB8DGGEFE@<@C@:>?,9,<?FFGCGGG?,<<E<,9F:<?@<AEG9A?CEF,AFGECF,@EEFEDF,A"[::-1]
 
-        mergedSequence, qualityScores, noOfCol = mergeUnpaired(left, right, lQuality, rQuality)
+        mergedSequence, qualityScores, noOfCol = self.readPairer.mergeUnpaired(left, right, lQuality, rQuality)
 
         expectedMerged = "TCTTTCTGCCTCATGCTCTCTCCAACAGGCTTGCAGCCAATTTACTGGAGCAGGGATGACGTAGCCCAGTGGCTCATGTGGGCTGAAAATGAGTTTTCTTTAAGGCCAATTGACAGCAACACATTTGAAATGAATGGCAAAGCTCTCCTGCTGCTGACCAAAGAGGACTTTCGCTATCGATCTCCTCATTCAGGTGAGAGTCTGGACTCTTGGCATATGCCCAACTTGGAAAATCTCTTAGTTAGTGGTTGGTCTTTAACACCCC"
         expecedQuality = ".............,.,...................,..,.........,.....,,.,..,..,,.....,.....?..........,.,.,.,,,......................,,.?..............,.......,.,,...,?.........,?,?.,,...,.......,...?,.?,?,,??.?,....,,.?..?,....,...?......,???....?.......,.........,??.,,....,..,?"
@@ -199,7 +202,7 @@ class PairedFASTQAligher(unittest.TestCase):
         lQuality = "CCCCCGEEFGGGGGGGGEGCCEFGCEGGFCFGGGGDG@EFGFGGGAFGGGEGGGGCDEEGFGGEEEFEG@CFEDEEGCFGCEFE@FEGFGG=4+,EFF7D=FGG=CEFB??,CFGFGG<,?,,4B+:,,,8<5,,C,B:,,+5,:,:,@B,"
         rQuality = "CCCCCFGGGGGGFGGFGFFGGGGGG;CEDEGGGGGGFGGGGGDG<AEFG?EBFFGGG:@CEGCF>FECFFFGGE<BFF,BCE,EABFB:>>FDFF9FF,=3@F<>:C*3<*5<****8,<;<,@,8;:=E*1*1*++2,27,,,***4*2+"[::-1]
 
-        mergedSequence, qualityScores, noOfCol = mergeUnpaired(left, right, lQuality, rQuality)
+        mergedSequence, qualityScores, noOfCol = self.readPairer.mergeUnpaired(left, right, lQuality, rQuality)
 
         expectedMerged = "TGCAGTCCCAGCCCACAGCCCCCCTCCTCCCTCAGACTCAGGAGTCCATA"
         expecedQuality = ".................................................."
